# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AURA AI Personal Assistant API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WeatherInfo {
  temp: number;
  condition: string;
  recommendation: string;
}

export interface NewsItem {
  id: number;
  title: string;
  summary: string;
  category: string;
  source: string;
  publishedAt: string;
  readTime: number;
  insight?: string;
}

export interface Briefing {
  date: string;
  greeting: string;
  weather: WeatherInfo;
  scheduleCount: number;
  tasksDueToday: number;
  tasksOverdue: number;
  urgentMailCount: number;
  topNews: NewsItem[];
  insight: string;
  pendingFollowUps: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface Task {
  id: number;
  title: string;
  description?: string;
  status: TaskStatus;
  priority: TaskPriority;
  dueDate?: string;
  category?: string;
  createdAt: string;
}

export type CreateTaskInputPriority =
  (typeof CreateTaskInputPriority)[keyof typeof CreateTaskInputPriority];

export const CreateTaskInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface CreateTaskInput {
  title: string;
  description?: string;
  priority: CreateTaskInputPriority;
  dueDate?: string;
  category?: string;
}

export type UpdateTaskInputStatus =
  (typeof UpdateTaskInputStatus)[keyof typeof UpdateTaskInputStatus];

export const UpdateTaskInputStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type UpdateTaskInputPriority =
  (typeof UpdateTaskInputPriority)[keyof typeof UpdateTaskInputPriority];

export const UpdateTaskInputPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  urgent: "urgent",
} as const;

export interface UpdateTaskInput {
  title?: string;
  description?: string;
  status?: UpdateTaskInputStatus;
  priority?: UpdateTaskInputPriority;
  dueDate?: string;
  category?: string;
}

export type MailPriority = (typeof MailPriority)[keyof typeof MailPriority];

export const MailPriority = {
  urgent: "urgent",
  important: "important",
  reference: "reference",
  spam: "spam",
} as const;

export interface Mail {
  id: number;
  sender: string;
  senderEmail: string;
  subject: string;
  preview: string;
  summary: string;
  priority: MailPriority;
  isRead: boolean;
  hasFollowUp: boolean;
  receivedAt: string;
}

export type ReplyDraftInputTone =
  (typeof ReplyDraftInputTone)[keyof typeof ReplyDraftInputTone];

export const ReplyDraftInputTone = {
  formal: "formal",
  friendly: "friendly",
  brief: "brief",
} as const;

export interface ReplyDraftInput {
  tone: ReplyDraftInputTone;
}

export interface ReplyDraft {
  subject: string;
  body: string;
  tone: string;
}

export type ReminderType = (typeof ReminderType)[keyof typeof ReminderType];

export const ReminderType = {
  time: "time",
  location: "location",
  habit: "habit",
  person: "person",
} as const;

export interface Reminder {
  id: number;
  title: string;
  type: ReminderType;
  scheduledAt?: string;
  isCompleted: boolean;
  repeatPattern?: string;
  relatedPerson?: string;
}

export type CreateReminderInputType =
  (typeof CreateReminderInputType)[keyof typeof CreateReminderInputType];

export const CreateReminderInputType = {
  time: "time",
  location: "location",
  habit: "habit",
  person: "person",
} as const;

export interface CreateReminderInput {
  title: string;
  type: CreateReminderInputType;
  scheduledAt?: string;
  repeatPattern?: string;
  relatedPerson?: string;
}

export interface ChatMessageInput {
  message: string;
  context?: string;
}

export interface QuickAction {
  label: string;
  action: string;
}

export interface ChatResponse {
  id: number;
  message: string;
  actions: QuickAction[];
  createdAt: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  message: string;
  createdAt: string;
}

export type GetTasksParams = {
  status?: GetTasksStatus;
};

export type GetTasksStatus =
  (typeof GetTasksStatus)[keyof typeof GetTasksStatus];

export const GetTasksStatus = {
  todo: "todo",
  in_progress: "in_progress",
  done: "done",
} as const;

export type GetMailsParams = {
  priority?: GetMailsPriority;
};

export type GetMailsPriority =
  (typeof GetMailsPriority)[keyof typeof GetMailsPriority];

export const GetMailsPriority = {
  urgent: "urgent",
  important: "important",
  reference: "reference",
  spam: "spam",
} as const;

export type GetNewsParams = {
  category?: string;
};