# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AURA AI Personal Assistant API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: briefing
    description: Morning briefing
  - name: tasks
    description: Task management
  - name: mail
    description: Smart mail management
  - name: news
    description: News and insights
  - name: reminders
    description: Smart reminders
  - name: chat
    description: AI chat interface
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /briefing:
    get:
      operationId: getBriefing
      tags: [briefing]
      summary: Get today's morning briefing
      responses:
        "200":
          description: Today's briefing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Briefing"

  /tasks:
    get:
      operationId: getTasks
      tags: [tasks]
      summary: Get all tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [todo, in_progress, done]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"

  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskInput"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /mail:
    get:
      operationId: getMails
      tags: [mail]
      summary: Get emails
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [urgent, important, reference, spam]
      responses:
        "200":
          description: List of emails
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mail"

  /mail/{id}/reply-draft:
    post:
      operationId: generateReplyDraft
      tags: [mail]
      summary: Generate reply draft for an email
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyDraftInput"
      responses:
        "200":
          description: Reply draft
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyDraft"

  /news:
    get:
      operationId: getNews
      tags: [news]
      summary: Get curated news
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: News articles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"

  /reminders:
    get:
      operationId: getReminders
      tags: [reminders]
      summary: Get all reminders
      responses:
        "200":
          description: List of reminders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reminder"
    post:
      operationId: createReminder
      tags: [reminders]
      summary: Create a reminder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReminderInput"
      responses:
        "201":
          description: Created reminder
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reminder"

  /chat:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to AURA AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"

  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history
      responses:
        "200":
          description: Chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Briefing:
      type: object
      properties:
        date:
          type: string
        greeting:
          type: string
        weather:
          $ref: "#/components/schemas/WeatherInfo"
        scheduleCount:
          type: integer
        tasksDueToday:
          type: integer
        tasksOverdue:
          type: integer
        urgentMailCount:
          type: integer
        topNews:
          type: array
          items:
            $ref: "#/components/schemas/NewsItem"
        insight:
          type: string
        pendingFollowUps:
          type: integer
      required:
        - date
        - greeting
        - weather
        - scheduleCount
        - tasksDueToday
        - tasksOverdue
        - urgentMailCount
        - topNews
        - insight
        - pendingFollowUps

    WeatherInfo:
      type: object
      properties:
        temp:
          type: integer
        condition:
          type: string
        recommendation:
          type: string
      required:
        - temp
        - condition
        - recommendation

    Task:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
        category:
          type: string
        createdAt:
          type: string
      required:
        - id
        - title
        - status
        - priority
        - createdAt

    CreateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
        category:
          type: string
      required:
        - title
        - priority

    UpdateTaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [todo, in_progress, done]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        dueDate:
          type: string
        category:
          type: string

    Mail:
      type: object
      properties:
        id:
          type: integer
        sender:
          type: string
        senderEmail:
          type: string
        subject:
          type: string
        preview:
          type: string
        summary:
          type: string
        priority:
          type: string
          enum: [urgent, important, reference, spam]
        isRead:
          type: boolean
        hasFollowUp:
          type: boolean
        receivedAt:
          type: string
      required:
        - id
        - sender
        - senderEmail
        - subject
        - preview
        - summary
        - priority
        - isRead
        - hasFollowUp
        - receivedAt

    ReplyDraftInput:
      type: object
      properties:
        tone:
          type: string
          enum: [formal, friendly, brief]
      required:
        - tone

    ReplyDraft:
      type: object
      properties:
        subject:
          type: string
        body:
          type: string
        tone:
          type: string
      required:
        - subject
        - body
        - tone

    NewsItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        readTime:
          type: integer
        insight:
          type: string
      required:
        - id
        - title
        - summary
        - category
        - source
        - publishedAt
        - readTime

    Reminder:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [time, location, habit, person]
        scheduledAt:
          type: string
        isCompleted:
          type: boolean
        repeatPattern:
          type: string
        relatedPerson:
          type: string
      required:
        - id
        - title
        - type
        - isCompleted

    CreateReminderInput:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [time, location, habit, person]
        scheduledAt:
          type: string
        repeatPattern:
          type: string
        relatedPerson:
          type: string
      required:
        - title
        - type

    ChatMessageInput:
      type: object
      properties:
        message:
          type: string
        context:
          type: string
      required:
        - message

    ChatResponse:
      type: object
      properties:
        id:
          type: integer
        message:
          type: string
        actions:
          type: array
          items:
            $ref: "#/components/schemas/QuickAction"
        createdAt:
          type: string
      required:
        - id
        - message
        - actions
        - createdAt

    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        message:
          type: string
        createdAt:
          type: string
      required:
        - id
        - role
        - message
        - createdAt

    QuickAction:
      type: object
      properties:
        label:
          type: string
        action:
          type: string
      required:
        - label
        - action