# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AURA AI Personal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { QuickAction } from "./quickAction";

export interface ChatResponse {
  id: number;
  message: string;
  actions: QuickAction[];
  createdAt: string;
}