# lib/api-zod/src/generated/types/reminder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AURA AI Personal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ReminderType } from "./reminderType";

export interface Reminder {
  id: number;
  title: string;
  type: ReminderType;
  scheduledAt?: string;
  isCompleted: boolean;
  repeatPattern?: string;
  relatedPerson?: string;
}