# lib/api-zod/src/generated/types/replyDraft.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AURA AI Personal Assistant API
 * OpenAPI spec version: 0.1.0
 */

export interface ReplyDraft {
  subject: string;
  body: string;
  tone: string;
}