# lib/api-zod/src/generated/types/replyDraftInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AURA AI Personal Assistant API
 * OpenAPI spec version: 0.1.0
 */
import type { ReplyDraftInputTone } from "./replyDraftInputTone";

export interface ReplyDraftInput {
  tone: ReplyDraftInputTone;
}